generate usable code
this commit is HUGE. it provides an improved java stdlib as well as generating usable code with labels for jumps. Note if labels are unused the code wont work. but just remove them and the code should run properly. Also the transpiler currently panics because it couldn't handle generating code. I need to figure that out